refactor: hardening jose utilities
