+++ packages/jose/src/assert.ts
@@ -32,3 +32,7 @@ export const isAsymmetricKeyPair = (value: unknown): value is AsymmetricKeyPair 
 export const isJWKKey = (value: unknown): value is JsonWebKey => {
     return typeof value === "object" && value !== null && "kty" in value && typeof (value as JsonWebKey).kty === "string"
 }
+
+export const isCryptoKey = (value: unknown): value is CryptoKey => {
+    return typeof value === "object" && value !== null && "type" in value && typeof (value as CryptoKey).type === "string"
+}
